% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropProbaTables.R
\name{cropProbaTables}
\alias{cropProbaTables}
\title{Crop probability tables.}
\usage{
cropProbaTables(
  onset,
  cessation,
  precip_daily,
  planting_dates,
  crop_water_requirements,
  crop_growing_season_lengths,
  minimum_fraction = 0.95
)
}
\arguments{
\item{onset}{A CDT stations data object containing the onset data in days.
Output from CDT menu \strong{Season Onset}, the file \strong{Onset_days.txt}}

\item{cessation}{A CDT stations data object containing the onset data in days.
Output from CDT menu \strong{Season Cessation}, the file \strong{Cessation_days.txt}}

\item{precip_daily}{A CDT stations data object containing the daily precipitation.}

\item{planting_dates}{A vector containing the planting dates, date format \code{"dd/mm"}.
Must be in increasing order, e.g: c("10/11", "20/11", "01/12", "10/12"). For southern hemisphere, 
the planting dates can continue to the next year, e.g: c("01/12", "10/12", "20/01", "01/02").}

\item{crop_water_requirements}{A vector containing the crop water requirement.}

\item{crop_growing_season_lengths}{A vector containing the crop growing season lengths.}

\item{minimum_fraction}{Minimum fraction of non-missing daily precipitation for each season.
If the fraction of non-missing data for one season is less than this number, 
the total of rainfall for that season will be missing.}
}
\value{
A list object
\itemize{
  \item{\strong{corp_proba}: }{
    A named list containing the crop probability. The names of the list are the growing season lengths.
    Each element of the list contains a \code{data.frame}, the 1st column indicates the growing season lengths,
    the 2nd column for the planting date, the 3rd column for the water requirements 
    and the next columns contain the probabilities for the stations
   }
  \item{\strong{stations}: }{
    A list containing the stations coordinates
    \itemize{
      \item{id: }{Vector of the stations id}
      \item{lon: }{Vector of the stations longitude}
      \item{lat: }{Vector of the stations latitude}
      \item{year: }{Vector of the start year for the season}
    }
  }
 \item{\strong{planting_data}: }{
   A named list containing the data for each planting date. The names of the list are the planting dates.
   Each element of the list contains a list of matrix of the amount of precipitation and the length of the season 
   starting from the planting date up to the cessation date. The row indicates the season and column the stations.
  }
 \item{\strong{seasonal_data}: }{
   A named list containing the seasonal data.
   \itemize{
     \item{season: }{Vector of the season}
     \item{origin: }{Vector origin date for the onset and cessation}
     \item{onset: }{Matrix of the onset data, row indicates the season and column the stations}
     \item{cessation: }{Matrix of the cessation data, same dimension as onset}
     \item{length: }{Matrix of the length of season, same dimension as onset}
     \item{amount: }{Matrix of the seasonal rainfall, same dimension as onset}
   }
  }
}
}
\description{
Compute the crop probability tables.
}
\examples{
\dontrun{

library(cropProbaTables)

# read daily precipitation data
precip_file <- "D:/DATA/DGM2021/precip_daily_1981-2020.csv"
precip_daily <- readCDTStationData(precip_file, sep = ",", missing = "-99")

# read onset data
onset_file <- "D:/DATA/DGM2021/ONSET_data/CDTSTATIONS/Onset_days.txt"
onset <- readCDTStationData(onset_file, sep = " ", missing = "-99")

# read cessation data
cessation_file <- "D:/DATA/DGM2021/CESSATION_data/CDTSTATIONS/Cessation_days.txt"
cessation <- readCDTStationData(cessation_file, sep = " ", missing = "-99")

out <- cropProbaTables(
            onset, cessation, precip_daily,
            planting_dates = c("10/11", "20/11", "01/12", "10/12"),
            crop_water_requirements = c(300, 400, 500),
            crop_growing_season_lengths = c(90, 100, 110, 120),
            minimum_fraction = 0.95
          )

output_directory <- "D:/DATA/DGM2021/CROP_PROBA"
writeProbaTables(out$corp_proba, output_directory)
}

}
