% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropProbaTables.R
\name{writeProbaTables}
\alias{writeProbaTables}
\title{Write crop probability tables.}
\usage{
writeProbaTables(crop_proba, output_directory)
}
\arguments{
\item{crop_proba}{The crop probability computed from \code{cropProbaTables}.}

\item{output_directory}{Folder to save the output.}
}
\description{
Write crop probability tables to CSV file.
}
