% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropProbaTables.R
\name{writeSeasonProba}
\alias{writeSeasonProba}
\title{Write rainy season start probability.}
\usage{
writeSeasonProba(season_proba, output_file)
}
\arguments{
\item{season_proba}{The rainy season start probability computed from \code{cropProbaTables}.}

\item{output_file}{The full path to the file to save the table.}
}
\description{
Write rainy season start probability table to CSV file.
}
